finalizado!

+++ src/app/pages/clients/list-clients/list-clients.component.ts
@@ -1,15 +1,15 @@
-import {Component, EventEmitter, Input, OnInit, Output} from "@angular/core";
-import {Cliente} from "../../../core/models/cliente.model";
-import {ClienteService} from "../../../core/services/clienteService";
+import {Component, OnInit} from "@angular/core";
 import Swal from "sweetalert2";
-import {Tag} from "../../../core/models/tag.model";
-import {forkJoin} from "rxjs";
-import {TagService} from "../../../core/services/tagService";
-import {ClienteTag} from "../../../core/models/cliente-tag.model";
 import {Router} from "@angular/router";
+import {forkJoin, Subject} from "rxjs";
+import {Tag} from "../../../core/models/tag.model";
+import {Cliente} from "../../../core/models/cliente.model";
+import {TagService} from "../../../core/services/tagService";
+import {debounceTime, distinctUntilChanged} from "rxjs/operators";
+import {ClienteTag} from "../../../core/models/cliente-tag.model";
+import {ClienteService} from "../../../core/services/clienteService";
 
 type Action = "DELETE" | "EDIT" | "TOGGLE";
-
 
 @Component({
     selector: 'app-list-clients',
@@ -20,13 +20,12 @@ export class ListClientsComponent implements OnInit {
 
     loading = false;
     clientes: Cliente[] = [];
+    clienteSearch: Cliente[] = [];
     tags: Tag[] = [];
+    search: string = '';
+    sortDirection = true;
 
-    @Input() option?: number;
-    @Input() uuid?: string;
-
-    @Output() optionChange = new EventEmitter<number>();
-    @Output() uuidChange = new EventEmitter<string>();
+    searchSubject: Subject<string> = new Subject();
 
     constructor(
         private router: Router,
@@ -40,6 +39,14 @@ export class ListClientsComponent implements OnInit {
         this.loadAll(() => {
             this.loading = false;
         });
+
+        this.searchSubject.pipe(
+            debounceTime(300),
+            distinctUntilChanged()
+        ).subscribe(search => {
+            this.loading = true;
+            this.searchClients(search);
+        });
     }
 
     loadAll(callback: () => void): void {
@@ -50,6 +57,7 @@ export class ListClientsComponent implements OnInit {
             next: (result) => {
                 if (result.clientes.status === 'SUCCESS') {
                     this.clientes = result.clientes.data;
+                    this.clienteSearch = result.clientes.data;
                 }
                 if (result.tags.status === 'SUCCESS') {
                     this.tags = result.tags.data;
@@ -57,10 +65,40 @@ export class ListClientsComponent implements OnInit {
                 callback();
             },
             error: (error) => {
-                console.error('Erro ao carregar clientes e tags', error);
                 Swal.fire('Erro', 'Erro ao carregar clientes e tags', 'error');
                 callback();
             }
+        });
+    }
+
+    onSearch() {
+        if (this.search.length > 3) {
+            this.searchSubject.next(this.search);
+        } else {
+            this.clientes = this.clienteSearch;
+        }
+    }
+
+    searchClients(search: string) {
+        this.clienteService.searchClients(search).subscribe((response) => {
+            if (response.data) this.clientes = response.data;
+            this.loading = false;
+        });
+    }
+
+    sortClients(column: keyof Cliente): void {
+        this.sortDirection = !this.sortDirection;
+
+        this.clientes.sort((a, b) => {
+            const aValue = a[column];
+            const bValue = b[column];
+
+            if (aValue !== undefined && bValue !== undefined) {
+                if (aValue < bValue) return this.sortDirection ? -1 : 1;
+                if (aValue > bValue) return this.sortDirection ? 1 : -1;
+            }
+
+            return 0;
         });
     }
 
@@ -78,13 +116,22 @@ export class ListClientsComponent implements OnInit {
     performAction(action: Action, element?: Cliente, extra?: any) {
         switch (action) {
             case "DELETE": {
-                this.clienteService.deleteCliente(element?.uuid).subscribe((response) => {
-                    Swal.fire('Sucesso', 'Cliente deletado com sucesso', 'success').then(() => {
-                        this.loadAll(() => {
-                            this.loading = false
+                Swal.fire({
+                    "title": "Deseja realmente excluir o cliente?", "icon": "warning",
+                    "showCancelButton": true, "confirmButtonText": "Sim", "cancelButtonText": "Não"
+                })
+                    .then((result) => {
+                        if (result.isDismissed) {
+                            return;
+                        }
+                        this.clienteService.deleteCliente(element?.uuid).subscribe((response) => {
+                            Swal.fire('Sucesso', 'Cliente deletado com sucesso', 'success').then(() => {
+                                this.loadAll(() => {
+                                    this.loading = false
+                                });
+                            });
                         });
                     });
-                });
                 break;
             }
             case "EDIT": {
@@ -104,7 +151,6 @@ export class ListClientsComponent implements OnInit {
                             title: `Cliente ${extra ? 'ativado' : 'inativado'} com sucesso!`,
                             icon: 'success'
                         });
-                        this.optionChange.emit(1);
                     },
                     ({error}) => {
                         Swal.fire({"title": error, "icon": "error"});
